requirements.txt
python-docx
pdfminer.six
pytesseract
pillow
pypandoc
jinja2
pyyaml
latexcodec
latexmk
